Move Table component into Logs directory
