+++ src/components/k8sLogs/Table.jsx
